pyproject.toml
[project]
name = "readers-quest-backend"
version = "0.1.0"
requires-python = "==3.12.*"
description = "Grade pronunciation of words."
readme = "README.md"
license = "BSD-3-Clause"
license-files = ["LICENSE.md"]
dependencies = [
  "fastapi[standard]>=0.115.11",
  "librosa>=0.11.0",
  "numpy>=2.1.3",
  "openai>=1.66.3",
  "phonemizer>=3.3.0",
  "pydub>=0.25.1",
  "python-jose>=3.4.0",
  "scikit-learn==1.6.0",
  "torch==2.5.*",
  "transformers>=4.49.0",
]

[project.urls]
Repository = "https://github.com/byytelope/readers-quest-backend.git"

[tool.basedpyright]
reportCallInDefaultInitializer = false
reportMissingTypeStubs = false
reportUnknownVariableType = false
reportAny = false
